Add chat order management and improve error handling
This commit introduces a new structure for managing chat orders, including incoming and submitted requests.
It adds the `IncomingRequests` and `SubmittedRequests` components to handle user conversations more effectively.
Error handling has been improved across various components to provide better feedback to users, especially regarding network issues.

Additionally, the old `OrderPage` and related widgets have been removed to streamline the codebase.
The new implementation enhances user experience by providing a clearer interface for managing chat orders.
